refactor(apidom-ls): rename externalDocs to externalDocumentation
