+++ packages/apidom-ls/src/config/openapi/external-documentation/documentation.ts
