+++ packages/apidom-ls/src/config/openapi/external-documentation/meta.ts
